weatherproject/requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Django==4.0.7
django-celery-beat==2.3.0
django-filter==22.1
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dnspython==2.2.1
eventlet==0.33.1
greenlet==1.1.3
kombu==5.2.4
Markdown==3.4.1
prompt-toolkit==3.0.30
PyJWT==2.4.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
vine==5.0.0
wcwidth==0.2.5
